[GPT] read gpt header and partitions

+++ include/klib/data_structure/hash_table.h
@@ -23,7 +23,7 @@ void hash_init(hash_t* hash);
  * @param key 
  * @return list_node_t* 
  */
-list_node_t* hash_insert(hash_t* hash, int key, void *value);
+list_node_t* hash_insert(hash_t* hash, size_t key, void *value);
 
 /**
  * @brief 
@@ -32,4 +32,4 @@ list_node_t* hash_insert(hash_t* hash, int key, void *value);
  * @param key 
  * @return list_node_t* 
  */
-list_node_t* hash_lookup(hash_t* hash, int key);
+list_node_t* hash_lookup(hash_t* hash, size_t key);
