+++ src/klib/data_structure/hash_table.c
@@ -8,12 +8,12 @@ void hash_init(hash_t* hash)
     }
 }
 
-list_node_t* hash_insert(hash_t* hash, int key, void *value)
+list_node_t* hash_insert(hash_t* hash, size_t key, void *value)
 {
     return list_insert(&hash->lists[key % BUCKETS], key, value);
 }
 
-list_node_t* hash_lookup(hash_t* hash, int key)
+list_node_t* hash_lookup(hash_t* hash, size_t key)
 {
     return list_lookup(&hash->lists[key % BUCKETS], key);
 }
